+++ src/vt/utils/compress/decompressor.impl.h
@@ -5,7 +5,7 @@
 //                             decompressor.impl.h
 //                       DARMA/vt => Virtual Transport
 //
-// Copyright 2019-2021 National Technology & Engineering Solutions of Sandia, LLC
+// Copyright 2019-2024 National Technology & Engineering Solutions of Sandia, LLC
 // (NTESS). Under the terms of Contract DE-NA0003525 with NTESS, the U.S.
 // Government retains certain rights in this software.
 //
@@ -136,7 +136,7 @@ std::size_t Decompressor<Readable>::read(
       // we have hit an unknown error, print the code and corresponding message!
       auto error_code = BrotliDecoderGetErrorCode(dec_);
       auto error_str = fmt::format(
-        "code={}, msg={}\n", error_code, BrotliDecoderErrorString(error_code)
+        "code={}, msg={}\n", static_cast<int32_t>(error_code), BrotliDecoderErrorString(error_code)
       );
       vtAbort(error_str);
       break;
